Features: - Added customisable loginUrl - Removed width definitions to allow consuming apps handle layout - Refactored stories for clearer identification of roles - Updated references to types
