+++ components/x-privacy-manager/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "noEmit": true,
+    "outDir": "dist/types",
     "allowJs": true,
     "target": "es2015",
     "module": "commonjs",
@@ -8,6 +8,6 @@
     "noImplicitAny": true,
     "jsx": "react"
   },
-  "include": ["src/**/*.js", "src/**/*.jsx", "typings/x-privacy-manager.d.ts"],
+  "include": ["src/**/*.js", "src/**/*.jsx", "typings/*.d.ts"],
   "exclude": ["node_modules"]
 }
